+++ CHANGELOG.rst
@@ -11,7 +11,14 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 **Added**
 
+* Check for non-deterministic functions for mlflow-tensorflow linter
+* Check for all_reduce for mlflow-xgboost templates
+* Check for OS for system-intelligence runs. If not Linux -> don't run system-intelligence
+* .gitattributes to templates, which ignores mlruns
+
 **Fixed**
+
+* Sync now operates correctly with the correct PR URL
 
 **Dependencies**
 
@@ -24,6 +31,8 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 **Added**
 
 **Fixed**
+
+* Conda report generation
 
 **Dependencies**